Apply selected layout on mount
This mainly comes up in dev, where the dev server hot reloads the app
without resetting the forms, etc., that are feeding the Graph component
props. Previously, regardless of the prop value for layout, the graph
would apply the force based layout; now it applies whatever layout is
actually the prop value.